+++ gammapy/catalog/gammacat.py
@@ -12,9 +12,15 @@ import numpy as np
 from astropy import units as u
 from astropy.table import Table
 from gammapy.cube.models import SkyModel, SkyModels
-from gammapy.modeling.models import SkyGaussian, SkyPointSource, SkyShell
+from gammapy.modeling.models import (
+    ExponentialCutoffPowerLaw,
+    PowerLaw,
+    PowerLaw2,
+    SkyGaussian,
+    SkyPointSource,
+    SkyShell,
+)
 from gammapy.spectrum import FluxPoints
-from gammapy.spectrum.models import ExponentialCutoffPowerLaw, PowerLaw, PowerLaw2
 from gammapy.utils.scripts import make_path
 from .core import SourceCatalog, SourceCatalogObject
