+++ gammapy/cube/exposure.py
@@ -1,7 +1,7 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 import numpy as np
 from gammapy.maps import WcsNDMap
-from gammapy.spectrum.models import PowerLaw
+from gammapy.modeling.models import PowerLaw
 
 __all__ = ["make_map_exposure_true_energy"]
