+++ gammapy/cube/models.py
@@ -4,8 +4,7 @@ from pathlib import Path
 import numpy as np
 import astropy.units as u
 from gammapy.maps import Map
-from gammapy.modeling.models import SkySpatialModel
-from gammapy.spectrum.models import SpectralModel
+from gammapy.modeling.models import SkySpatialModel, SpectralModel
 from gammapy.utils.fitting import Model, Parameter, Parameters
 from gammapy.utils.scripts import make_path, write_yaml
