+++ gammapy/cube/tests/test_fit.py
@@ -13,8 +13,7 @@ from gammapy.irf import (
     EnergyDispersion,
 )
 from gammapy.maps import Map, MapAxis, WcsGeom
-from gammapy.modeling.models import SkyGaussian
-from gammapy.spectrum.models import PowerLaw
+from gammapy.modeling.models import PowerLaw, SkyGaussian
 from gammapy.utils.fitting import Fit
 from gammapy.utils.testing import mpl_plot_check, requires_data, requires_dependency
