+++ gammapy/irf/effective_area.py
@@ -276,7 +276,7 @@ class EffectiveAreaTable:
         energy : `~astropy.units.Quantity`
             Energy corresponding to the given aeff.
         """
-        from gammapy.spectrum.models import TableModel
+        from gammapy.modeling.models import TableModel
 
         energy = self.energy.center
