+++ gammapy/irf/energy_dispersion.py
@@ -477,7 +477,7 @@ class EnergyDispersion:
         bias_energy : `~astropy.unit.Quantity`
             Reconstructed energy corresponding to the given bias.
         """
-        from gammapy.spectrum.models import TableModel
+        from gammapy.modeling.models import TableModel
 
         e_true = self.e_true.center
         values = self.get_bias(e_true)
