+++ gammapy/modeling/models/spectrum/__init__.py
@@ -1,0 +1,5 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""Spectral models."""
+from .core import *
+from .cosmic_ray import *
+from .crab import *
