+++ gammapy/modeling/models/spectrum/core.py
@@ -6,11 +6,11 @@ import scipy.optimize
 import scipy.special
 import astropy.units as u
 from astropy.table import Table
+from gammapy.spectrum.utils import integrate_spectrum
 from gammapy.utils.energy import energy_logspace
 from gammapy.utils.fitting import Model, Parameter, Parameters
 from gammapy.utils.interpolation import ScaledRegularGridInterpolator
 from gammapy.utils.scripts import make_path
-from .utils import integrate_spectrum
 
 __all__ = [
     "SpectralModel",
