+++ gammapy/spectrum/__init__.py
@@ -2,8 +2,6 @@
 """1D spectrum analysis."""
 from .background_estimate import *
 from .core import *
-from .cosmic_ray import *
-from .crab import *
 from .dataset import *
 from .extract import *
 from .flux_point import *
