+++ gammapy/spectrum/tests/test_dataset.py
@@ -7,13 +7,13 @@ from astropy.table import Table
 from astropy.time import Time
 from gammapy.data import GTI
 from gammapy.irf import EffectiveAreaTable, EnergyDispersion
+from gammapy.modeling.models import ConstantModel, ExponentialCutoffPowerLaw, PowerLaw
 from gammapy.spectrum import (
     CountsSpectrum,
     SpectrumDataset,
     SpectrumDatasetOnOff,
     SpectrumDatasetOnOffStacker,
 )
-from gammapy.spectrum.models import ConstantModel, ExponentialCutoffPowerLaw, PowerLaw
 from gammapy.utils.fitting import Fit
 from gammapy.utils.random import get_random_state
 from gammapy.utils.testing import mpl_plot_check, requires_data, requires_dependency
