+++ gammapy/spectrum/tests/test_flux_point.py
@@ -5,8 +5,8 @@ from numpy.testing import assert_allclose
 import astropy.units as u
 from astropy.table import Table
 from gammapy.catalog.fermi import SourceCatalog3FGL
+from gammapy.modeling.models import PowerLaw, SpectralModel
 from gammapy.spectrum import FluxPoints, FluxPointsDataset
-from gammapy.spectrum.models import PowerLaw, SpectralModel
 from gammapy.utils.fitting import Fit
 from gammapy.utils.testing import (
     assert_quantity_allclose,
