+++ gammapy/utils/random/tests/test_inverse_cdf.py
@@ -9,8 +9,7 @@ from astropy.time import Time
 from gammapy.cube import MapEvaluator
 from gammapy.cube.models import SkyModel
 from gammapy.maps import Map, MapAxis
-from gammapy.modeling.models import LightCurveTableModel, SkyGaussian
-from gammapy.spectrum.models import PowerLaw
+from gammapy.modeling.models import LightCurveTableModel, PowerLaw, SkyGaussian
 from gammapy.utils.random import InverseCDFSampler, MapEventSampler
