+++ gammapy/utils/serialization/tests/make_test_data.py
@@ -11,8 +11,12 @@ from gammapy.cube.models import BackgroundModel, SkyDiffuseCube, SkyModel, SkyMo
 from gammapy.data import DataStore
 from gammapy.irf import make_mean_edisp, make_mean_psf
 from gammapy.maps import MapAxis, WcsGeom
-from gammapy.modeling.models import SkyGaussian, SkyPointSource
-from gammapy.spectrum.models import ExponentialCutoffPowerLaw, PowerLaw
+from gammapy.modeling.models import (
+    ExponentialCutoffPowerLaw,
+    PowerLaw,
+    SkyGaussian,
+    SkyPointSource,
+)
 from gammapy.utils.fitting import Datasets
 
 DATA_PATH = Path("gammapy/utils/serialization/tests/data")
