+++ gammapy/utils/serialization/xml.py
@@ -13,7 +13,7 @@ from gammapy.cube.models import SkyModel, SkyModels
 from gammapy.extern import xmltodict
 from gammapy.maps import Map
 from gammapy.modeling.models import image as spatial
-from gammapy.spectrum import models as spectral
+from gammapy.modeling.models import spectrum as spectral
 from gammapy.utils.fitting import Parameter, Parameters
 
 log = logging.getLogger(__name__)
